refactor nuxt 3 mock
