+++ packages/nuxt3/src/app/composables/component.ts
@@ -2,7 +2,7 @@ import { toRefs } from '@vue/reactivity'
 import { defineComponent, getCurrentInstance } from 'vue'
 import type { DefineComponent } from 'vue'
 import { useRoute } from 'vue-router'
-import type { LegacyContext } from '../legacy'
+import type { LegacyContext } from '../compat/legacy-app'
 import { useNuxtApp } from '../nuxt'
 import { useAsyncData } from './asyncData'
