+++ packages/nuxt3/src/app/nuxt.ts
@@ -3,7 +3,7 @@ import { getCurrentInstance, reactive } from 'vue'
 import type { App, VNode } from 'vue'
 import { createHooks, Hookable } from 'hookable'
 import type { RuntimeConfig } from '@nuxt/kit'
-import { legacyPlugin, LegacyContext } from './legacy'
+import { legacyPlugin, LegacyContext } from './compat/legacy-app'
 
 type NuxtMeta = {
   htmlAttrs?: string
